Add Docker Support
This change adds a Dockerfile to support local development and
production releases in containerized environments. The change itself
include documentation on how to start the docker container locally as
well as information on running the python application natively.

Support for the python module xattr was removed due to it not being used
and its lack of cross-platform support.